Move `CellSwitch` to the `widget` package
