+++ android/src/main/res/layout/split_tunnelling_header.xml
@@ -36,10 +36,10 @@
                   android:textSize="20sp"
                   android:textStyle="bold"
                   android:text="@string/enabled" />
-        <net.mullvad.mullvadvpn.ui.CellSwitch android:id="@+id/enabled_toggle"
-                                              android:layout_width="52dp"
-                                              android:layout_height="32dp"
-                                              android:layout_weight="0" />
+        <net.mullvad.mullvadvpn.ui.widget.CellSwitch android:id="@+id/enabled_toggle"
+                                                     android:layout_width="52dp"
+                                                     android:layout_height="32dp"
+                                                     android:layout_weight="0" />
     </LinearLayout>
     <LinearLayout android:id="@+id/exclude_applications"
                   android:layout_width="match_parent"
